requirements.txt
atpublic==4.0
attrs==23.1.0
Babel==2.8.0
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==2.3.3
Flask-Mail==0.9.1
Flask-Session==0.5.0
flask-session2==1.3.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pycairo==1.25.1
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-http-client==3.3.7
pytz==2022.7.1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
sendgrid==6.11.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==2.0.23
starkbank-ecdsa==2.2.0
systemd-python==234
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
wadllib==1.3.6
Werkzeug==3.0.1
zipp==1.0.0
